+++ EmergencyDataCollectionApp/App.js
@@ -1,11 +1,11 @@
 import { StatusBar } from 'expo-status-bar';
 import { SafeAreaView, StyleSheet, Text, View } from 'react-native';
 import NavigationBar from './components/navigation/NavigationBar';
+import { NavigationContainer } from '@react-navigation/native';
 
 export default function App() {
   return (
     <SafeAreaView>
-      <View style={{height: '21%'}}></View>
       <NavigationBar></NavigationBar>
       <View style={styles.container}>
         <Text>Hello SER401!</Text>
