+++ EmergencyDataCollectionApp/package.json
@@ -9,10 +9,17 @@
     "web": "expo start --web"
   },
   "dependencies": {
-    "expo": "~49.0.13",
+    "@react-navigation/bottom-tabs": "^6.5.11",
+    "@react-navigation/material-top-tabs": "^6.6.5",
+    "@react-navigation/native": "^6.1.9",
+    "expo": "^49.0.16",
     "expo-status-bar": "~1.6.0",
     "react": "18.2.0",
-    "react-native": "0.72.6"
+    "react-native": "0.72.6",
+    "react-native-pager-view": "^6.2.0",
+    "react-native-safe-area-context": "4.6.3",
+    "react-native-screens": "~3.22.0",
+    "react-native-tab-view": "^3.5.2"
   },
   "devDependencies": {
     "@babel/core": "^7.20.0"
